IPlayer: change GetTotalTime() to return milliseconds in an int64_t

+++ xbmc/cores/ExternalPlayer/ExternalPlayer.h
@@ -72,7 +72,7 @@ public:
 
   virtual void SeekTime(int64_t iTime);
   virtual int64_t GetTime();
-  virtual int GetTotalTime();
+  virtual int64_t GetTotalTime();
   virtual void ToFFRW(int iSpeed);
   virtual void ShowOSD(bool bOnoff);
   virtual void DoAudioWork()                                    {}
